+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.18.0]
+
+### Added
+
+- Toggle to show rider or emergency contact number
+
 ## [0.17.0]
 
 ### Changed